Make the preprocessors vector a field in the ReplicaFactory class
