+++ bftengine/include/bftengine/Replica.hpp
@@ -90,33 +90,15 @@ class IControlHandler {
   virtual ~IControlHandler() = default;
 };
 
+class IExternalObject {
+ public:
+  virtual ~IExternalObject() = default;
+  virtual void setAggregator(const std::shared_ptr<concordMetrics::Aggregator> &a) = 0;
+  virtual void stop() = 0;
+};
+
 class IReplica {
  public:
-  using IReplicaPtr = std::unique_ptr<IReplica>;
-  static IReplicaPtr createNewReplica(const ReplicaConfig &,
-                                      std::shared_ptr<IRequestsHandler>,
-                                      IStateTransfer *,
-                                      bft::communication::ICommunication *,
-                                      MetadataStorage *,
-                                      std::shared_ptr<concord::performance::PerformanceManager> pm,
-                                      const std::shared_ptr<concord::secretsmanager::ISecretsManagerImpl> &sm,
-                                      const std::function<void(bool)> &);
-  static IReplicaPtr createNewReplica(const ReplicaConfig &,
-                                      std::shared_ptr<IRequestsHandler>,
-                                      IStateTransfer *,
-                                      bft::communication::ICommunication *,
-                                      MetadataStorage *,
-                                      bool &erasedMetadata,
-                                      std::shared_ptr<concord::performance::PerformanceManager> pm,
-                                      const std::shared_ptr<concord::secretsmanager::ISecretsManagerImpl> &sm,
-                                      const std::function<void(bool)> &);
-
-  static IReplicaPtr createNewRoReplica(const ReplicaConfig &,
-                                        std::shared_ptr<IRequestsHandler>,
-                                        IStateTransfer *,
-                                        bft::communication::ICommunication *,
-                                        MetadataStorage *);
-
   virtual ~IReplica() = default;
 
   virtual bool isRunning() const = 0;
