+++ kvbc/include/ClientImp.h
@@ -40,7 +40,8 @@ class ClientImp : public IClient {
 
  protected:
   ClientImp() = default;
-  ~ClientImp() override = default;
+  ~ClientImp() override { delete comm_; };
+
 
   ClientConfig config_;
   std::unique_ptr<bftEngine::SeqNumberGeneratorForClientRequests> seqGen_;
