+++ kvbc/include/Replica.h
@@ -36,6 +36,7 @@
 #include "AdaptivePruningManager.hpp"
 #include "kvbc_app_filter/value_from_kvbc_proto.h"
 #include "newest_public_event_group_record_time.h"
+#include "ReplicaFactory.hpp"
 
 namespace concord::kvbc {
 
@@ -246,7 +247,7 @@ class Replica : public IReplica,
   std::shared_ptr<storage::IDBClient> m_metadataDBClient;
   std::unique_ptr<bft::communication::ICommunication> m_ptrComm;
   const bftEngine::ReplicaConfig &replicaConfig_;
-  bftEngine::IReplica::IReplicaPtr m_replicaPtr = nullptr;
+  bftEngine::ReplicaFactory::IReplicaPtr m_replicaPtr;
   std::shared_ptr<ICommandsHandler> m_cmdHandler = nullptr;
   bftEngine::IStateTransfer *m_stateTransfer = nullptr;
   concord::storage::DBMetadataStorage *m_metadataStorage = nullptr;
@@ -263,6 +264,7 @@ class Replica : public IReplica,
   ReplicaResourceEntity replicaResources_;
   concord::util::ThreadPool blocks_io_workers_pool;
   performance::AdaptivePruningManager AdaptivePruningManager_;
+
   struct Recorders {
     static constexpr uint64_t MAX_VALUE_MICROSECONDS = 2ULL * 1000ULL * 1000ULL;  // 2 seconds
     const std::string component_ = "iappstate";
